Add a tip to a phylogenetic tree

AddTip adds a tip to a phylogenetic tree at a specified location.

AddTip(tree, where = sample.int(tree$Nnode * 2 + 2L, size = 1) - 1L,
  label = "New tip")

AddTipEverywhere(tree, label = "New tip", includeRoot = FALSE)

Arguments

tree
A tree of class phylo.

where
The node or tip that should form the sister taxon to the new node. To add a new tip at the root, use where = 0. By default, the new tip is added to a random edge.

label
Character string providing the label to apply to the new tip.

includeRoot
Logical; if TRUE, the three positions adjacent to the root edge are considered to represent distinct edges.

Value

AddTip returns a tree of class phylo with an additional tip at the desired location.

AddTipEverywhere returns a list of class multiPhylo containing the trees produced by adding label to each edge of tree in turn.

Details

AddTip extends bind.tree, which cannot handle single-taxon trees.

Functions

AddTipEverywhere: AddTipEverywhere Add a tip to each edge in turn

See also

Examples

{
  library('ape')
  plot(tree <- rtree(10, br=NULL)); nodelabels(); nodelabels(15, 15, bg='green'); dev.new()
  plot(AddTip(tree, 15, 'NEW_TIP'))
}

#> dev.new(): using pdf(file="Rplots245.pdf")

backbone <- BalancedTree(4)
additions <- AddTipEverywhere(backbone, includeRoot = TRUE)
par(mfrow=c(2, 4), mar=rep(0.3, 4), cex=0.9)
lapply(additions, plot)

#> [[1]]
#> [[1]]$type
#> [1] "phylogram"
#> 
#> [[1]]$use.edge.length
#> [1] FALSE
#> 
#> [[1]]$node.pos
#> [1] 1
#> 
#> [[1]]$node.depth
#> [1] 1
#> 
#> [[1]]$show.tip.label
#> [1] TRUE
#> 
#> [[1]]$show.node.label
#> [1] FALSE
#> 
#> [[1]]$font
#> [1] 3
#> 
#> [[1]]$cex
#> [1] 0.9
#> 
#> [[1]]$adj
#> [1] 0
#> 
#> [[1]]$srt
#> [1] 0
#> 
#> [[1]]$no.margin
#> [1] FALSE
#> 
#> [[1]]$label.offset
#> [1] 0
#> 
#> [[1]]$x.lim
#> [1] 0.000000 7.155143
#> 
#> [[1]]$y.lim
#> [1] 1 5
#> 
#> [[1]]$direction
#> [1] "rightwards"
#> 
#> [[1]]$tip.color
#> [1] "black"
#> 
#> [[1]]$Ntip
#> [1] 5
#> 
#> [[1]]$Nnode
#> [1] 4
#> 
#> [[1]]$root.time
#> NULL
#> 
#> [[1]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[2]]
#> [[2]]$type
#> [1] "phylogram"
#> 
#> [[2]]$use.edge.length
#> [1] FALSE
#> 
#> [[2]]$node.pos
#> [1] 1
#> 
#> [[2]]$node.depth
#> [1] 1
#> 
#> [[2]]$show.tip.label
#> [1] TRUE
#> 
#> [[2]]$show.node.label
#> [1] FALSE
#> 
#> [[2]]$font
#> [1] 3
#> 
#> [[2]]$cex
#> [1] 0.9
#> 
#> [[2]]$adj
#> [1] 0
#> 
#> [[2]]$srt
#> [1] 0
#> 
#> [[2]]$no.margin
#> [1] FALSE
#> 
#> [[2]]$label.offset
#> [1] 0
#> 
#> [[2]]$x.lim
#> [1] 0.000000 7.234228
#> 
#> [[2]]$y.lim
#> [1] 1 5
#> 
#> [[2]]$direction
#> [1] "rightwards"
#> 
#> [[2]]$tip.color
#> [1] "black"
#> 
#> [[2]]$Ntip
#> [1] 5
#> 
#> [[2]]$Nnode
#> [1] 4
#> 
#> [[2]]$root.time
#> NULL
#> 
#> [[2]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[3]]
#> [[3]]$type
#> [1] "phylogram"
#> 
#> [[3]]$use.edge.length
#> [1] FALSE
#> 
#> [[3]]$node.pos
#> [1] 1
#> 
#> [[3]]$node.depth
#> [1] 1
#> 
#> [[3]]$show.tip.label
#> [1] TRUE
#> 
#> [[3]]$show.node.label
#> [1] FALSE
#> 
#> [[3]]$font
#> [1] 3
#> 
#> [[3]]$cex
#> [1] 0.9
#> 
#> [[3]]$adj
#> [1] 0
#> 
#> [[3]]$srt
#> [1] 0
#> 
#> [[3]]$no.margin
#> [1] FALSE
#> 
#> [[3]]$label.offset
#> [1] 0
#> 
#> [[3]]$x.lim
#> [1] 0.000000 7.234228
#> 
#> [[3]]$y.lim
#> [1] 1 5
#> 
#> [[3]]$direction
#> [1] "rightwards"
#> 
#> [[3]]$tip.color
#> [1] "black"
#> 
#> [[3]]$Ntip
#> [1] 5
#> 
#> [[3]]$Nnode
#> [1] 4
#> 
#> [[3]]$root.time
#> NULL
#> 
#> [[3]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[4]]
#> [[4]]$type
#> [1] "phylogram"
#> 
#> [[4]]$use.edge.length
#> [1] FALSE
#> 
#> [[4]]$node.pos
#> [1] 1
#> 
#> [[4]]$node.depth
#> [1] 1
#> 
#> [[4]]$show.tip.label
#> [1] TRUE
#> 
#> [[4]]$show.node.label
#> [1] FALSE
#> 
#> [[4]]$font
#> [1] 3
#> 
#> [[4]]$cex
#> [1] 0.9
#> 
#> [[4]]$adj
#> [1] 0
#> 
#> [[4]]$srt
#> [1] 0
#> 
#> [[4]]$no.margin
#> [1] FALSE
#> 
#> [[4]]$label.offset
#> [1] 0
#> 
#> [[4]]$x.lim
#> [1] 0.000000 7.234228
#> 
#> [[4]]$y.lim
#> [1] 1 5
#> 
#> [[4]]$direction
#> [1] "rightwards"
#> 
#> [[4]]$tip.color
#> [1] "black"
#> 
#> [[4]]$Ntip
#> [1] 5
#> 
#> [[4]]$Nnode
#> [1] 4
#> 
#> [[4]]$root.time
#> NULL
#> 
#> [[4]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[5]]
#> [[5]]$type
#> [1] "phylogram"
#> 
#> [[5]]$use.edge.length
#> [1] FALSE
#> 
#> [[5]]$node.pos
#> [1] 1
#> 
#> [[5]]$node.depth
#> [1] 1
#> 
#> [[5]]$show.tip.label
#> [1] TRUE
#> 
#> [[5]]$show.node.label
#> [1] FALSE
#> 
#> [[5]]$font
#> [1] 3
#> 
#> [[5]]$cex
#> [1] 0.9
#> 
#> [[5]]$adj
#> [1] 0
#> 
#> [[5]]$srt
#> [1] 0
#> 
#> [[5]]$no.margin
#> [1] FALSE
#> 
#> [[5]]$label.offset
#> [1] 0
#> 
#> [[5]]$x.lim
#> [1] 0.000000 7.234228
#> 
#> [[5]]$y.lim
#> [1] 1 5
#> 
#> [[5]]$direction
#> [1] "rightwards"
#> 
#> [[5]]$tip.color
#> [1] "black"
#> 
#> [[5]]$Ntip
#> [1] 5
#> 
#> [[5]]$Nnode
#> [1] 4
#> 
#> [[5]]$root.time
#> NULL
#> 
#> [[5]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[6]]
#> [[6]]$type
#> [1] "phylogram"
#> 
#> [[6]]$use.edge.length
#> [1] FALSE
#> 
#> [[6]]$node.pos
#> [1] 1
#> 
#> [[6]]$node.depth
#> [1] 1
#> 
#> [[6]]$show.tip.label
#> [1] TRUE
#> 
#> [[6]]$show.node.label
#> [1] FALSE
#> 
#> [[6]]$font
#> [1] 3
#> 
#> [[6]]$cex
#> [1] 0.9
#> 
#> [[6]]$adj
#> [1] 0
#> 
#> [[6]]$srt
#> [1] 0
#> 
#> [[6]]$no.margin
#> [1] FALSE
#> 
#> [[6]]$label.offset
#> [1] 0
#> 
#> [[6]]$x.lim
#> [1] 0.000000 7.234228
#> 
#> [[6]]$y.lim
#> [1] 1 5
#> 
#> [[6]]$direction
#> [1] "rightwards"
#> 
#> [[6]]$tip.color
#> [1] "black"
#> 
#> [[6]]$Ntip
#> [1] 5
#> 
#> [[6]]$Nnode
#> [1] 4
#> 
#> [[6]]$root.time
#> NULL
#> 
#> [[6]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[7]]
#> [[7]]$type
#> [1] "phylogram"
#> 
#> [[7]]$use.edge.length
#> [1] FALSE
#> 
#> [[7]]$node.pos
#> [1] 1
#> 
#> [[7]]$node.depth
#> [1] 1
#> 
#> [[7]]$show.tip.label
#> [1] TRUE
#> 
#> [[7]]$show.node.label
#> [1] FALSE
#> 
#> [[7]]$font
#> [1] 3
#> 
#> [[7]]$cex
#> [1] 0.9
#> 
#> [[7]]$adj
#> [1] 0
#> 
#> [[7]]$srt
#> [1] 0
#> 
#> [[7]]$no.margin
#> [1] FALSE
#> 
#> [[7]]$label.offset
#> [1] 0
#> 
#> [[7]]$x.lim
#> [1] 0.000000 7.234228
#> 
#> [[7]]$y.lim
#> [1] 1 5
#> 
#> [[7]]$direction
#> [1] "rightwards"
#> 
#> [[7]]$tip.color
#> [1] "black"
#> 
#> [[7]]$Ntip
#> [1] 5
#> 
#> [[7]]$Nnode
#> [1] 4
#> 
#> [[7]]$root.time
#> NULL
#> 
#> [[7]]$align.tip.label
#> [1] FALSE
#> 
#> 

additions <- AddTipEverywhere(backbone, includeRoot = FALSE)
par(mfrow=c(2, 3))
lapply(additions, plot)

#> [[1]]
#> [[1]]$type
#> [1] "phylogram"
#> 
#> [[1]]$use.edge.length
#> [1] FALSE
#> 
#> [[1]]$node.pos
#> [1] 1
#> 
#> [[1]]$node.depth
#> [1] 1
#> 
#> [[1]]$show.tip.label
#> [1] TRUE
#> 
#> [[1]]$show.node.label
#> [1] FALSE
#> 
#> [[1]]$font
#> [1] 3
#> 
#> [[1]]$cex
#> [1] 0.66
#> 
#> [[1]]$adj
#> [1] 0
#> 
#> [[1]]$srt
#> [1] 0
#> 
#> [[1]]$no.margin
#> [1] FALSE
#> 
#> [[1]]$label.offset
#> [1] 0
#> 
#> [[1]]$x.lim
#> [1] 0.000000 4.921637
#> 
#> [[1]]$y.lim
#> [1] 1 5
#> 
#> [[1]]$direction
#> [1] "rightwards"
#> 
#> [[1]]$tip.color
#> [1] "black"
#> 
#> [[1]]$Ntip
#> [1] 5
#> 
#> [[1]]$Nnode
#> [1] 4
#> 
#> [[1]]$root.time
#> NULL
#> 
#> [[1]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[2]]
#> [[2]]$type
#> [1] "phylogram"
#> 
#> [[2]]$use.edge.length
#> [1] FALSE
#> 
#> [[2]]$node.pos
#> [1] 1
#> 
#> [[2]]$node.depth
#> [1] 1
#> 
#> [[2]]$show.tip.label
#> [1] TRUE
#> 
#> [[2]]$show.node.label
#> [1] FALSE
#> 
#> [[2]]$font
#> [1] 3
#> 
#> [[2]]$cex
#> [1] 0.66
#> 
#> [[2]]$adj
#> [1] 0
#> 
#> [[2]]$srt
#> [1] 0
#> 
#> [[2]]$no.margin
#> [1] FALSE
#> 
#> [[2]]$label.offset
#> [1] 0
#> 
#> [[2]]$x.lim
#> [1] 0.000000 4.921637
#> 
#> [[2]]$y.lim
#> [1] 1 5
#> 
#> [[2]]$direction
#> [1] "rightwards"
#> 
#> [[2]]$tip.color
#> [1] "black"
#> 
#> [[2]]$Ntip
#> [1] 5
#> 
#> [[2]]$Nnode
#> [1] 4
#> 
#> [[2]]$root.time
#> NULL
#> 
#> [[2]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[3]]
#> [[3]]$type
#> [1] "phylogram"
#> 
#> [[3]]$use.edge.length
#> [1] FALSE
#> 
#> [[3]]$node.pos
#> [1] 1
#> 
#> [[3]]$node.depth
#> [1] 1
#> 
#> [[3]]$show.tip.label
#> [1] TRUE
#> 
#> [[3]]$show.node.label
#> [1] FALSE
#> 
#> [[3]]$font
#> [1] 3
#> 
#> [[3]]$cex
#> [1] 0.66
#> 
#> [[3]]$adj
#> [1] 0
#> 
#> [[3]]$srt
#> [1] 0
#> 
#> [[3]]$no.margin
#> [1] FALSE
#> 
#> [[3]]$label.offset
#> [1] 0
#> 
#> [[3]]$x.lim
#> [1] 0.000000 4.921637
#> 
#> [[3]]$y.lim
#> [1] 1 5
#> 
#> [[3]]$direction
#> [1] "rightwards"
#> 
#> [[3]]$tip.color
#> [1] "black"
#> 
#> [[3]]$Ntip
#> [1] 5
#> 
#> [[3]]$Nnode
#> [1] 4
#> 
#> [[3]]$root.time
#> NULL
#> 
#> [[3]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[4]]
#> [[4]]$type
#> [1] "phylogram"
#> 
#> [[4]]$use.edge.length
#> [1] FALSE
#> 
#> [[4]]$node.pos
#> [1] 1
#> 
#> [[4]]$node.depth
#> [1] 1
#> 
#> [[4]]$show.tip.label
#> [1] TRUE
#> 
#> [[4]]$show.node.label
#> [1] FALSE
#> 
#> [[4]]$font
#> [1] 3
#> 
#> [[4]]$cex
#> [1] 0.66
#> 
#> [[4]]$adj
#> [1] 0
#> 
#> [[4]]$srt
#> [1] 0
#> 
#> [[4]]$no.margin
#> [1] FALSE
#> 
#> [[4]]$label.offset
#> [1] 0
#> 
#> [[4]]$x.lim
#> [1] 0.000000 4.921637
#> 
#> [[4]]$y.lim
#> [1] 1 5
#> 
#> [[4]]$direction
#> [1] "rightwards"
#> 
#> [[4]]$tip.color
#> [1] "black"
#> 
#> [[4]]$Ntip
#> [1] 5
#> 
#> [[4]]$Nnode
#> [1] 4
#> 
#> [[4]]$root.time
#> NULL
#> 
#> [[4]]$align.tip.label
#> [1] FALSE
#> 
#> 
#> [[5]]
#> [[5]]$type
#> [1] "phylogram"
#> 
#> [[5]]$use.edge.length
#> [1] FALSE
#> 
#> [[5]]$node.pos
#> [1] 1
#> 
#> [[5]]$node.depth
#> [1] 1
#> 
#> [[5]]$show.tip.label
#> [1] TRUE
#> 
#> [[5]]$show.node.label
#> [1] FALSE
#> 
#> [[5]]$font
#> [1] 3
#> 
#> [[5]]$cex
#> [1] 0.66
#> 
#> [[5]]$adj
#> [1] 0
#> 
#> [[5]]$srt
#> [1] 0
#> 
#> [[5]]$no.margin
#> [1] FALSE
#> 
#> [[5]]$label.offset
#> [1] 0
#> 
#> [[5]]$x.lim
#> [1] 0.000000 4.921637
#> 
#> [[5]]$y.lim
#> [1] 1 5
#> 
#> [[5]]$direction
#> [1] "rightwards"
#> 
#> [[5]]$tip.color
#> [1] "black"
#> 
#> [[5]]$Ntip
#> [1] 5
#> 
#> [[5]]$Nnode
#> [1] 4
#> 
#> [[5]]$root.time
#> NULL
#> 
#> [[5]]$align.tip.label
#> [1] FALSE
#> 
#>