Generate a tree with a specified outgroup

Deprecated. This function will be removed in a future version of TreeTools. Use RootTree() instead.

Usage

EnforceOutgroup(tree, outgroup)

# S3 method for class 'phylo'
EnforceOutgroup(tree, outgroup)

# S3 method for class 'character'
EnforceOutgroup(tree, outgroup)

Arguments

tree

Either a tree of class phylo; or (for EnforceOutgroup()) a character vector listing the names of all the taxa in the tree, from which a random tree will be generated.

outgroup

Character vector containing the names of taxa to include in the outgroup.

Value

EnforceOutgroup() returned a tree of class phylo where all outgroup taxa are sister to all remaining taxa, without modifying the ingroup topology.

Details

Given a tree or a list of taxa, EnforceOutgroup() rearranged the ingroup and outgroup taxa such that the two are sister taxa across the root, without changing the relationships within the ingroup or within the outgroup.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)