Number of leaves in a phylogenetic tree

NTip() extends ape::Ntip() to handle objects of class Splits and list, and edge matrices (equivalent to tree$edge).

Usage

NTip(phy)

# Default S3 method
NTip(phy)

# S3 method for class 'Splits'
NTip(phy)

# S3 method for class 'list'
NTip(phy)

# S3 method for class 'phylo'
NTip(phy)

# S3 method for class 'multiPhylo'
NTip(phy)

# S3 method for class 'phyDat'
NTip(phy)

# S3 method for class 'matrix'
NTip(phy)

Arguments

phy

Object representing one or more phylogenetic trees.

Value

NTip() returns an integer specifying the number of tips in each object in phy.

See also