Numeric index of each node in a tree NodeNumbers() returns a sequence corresponding to the nodes in a tree

Numeric index of each node in a tree NodeNumbers() returns a sequence corresponding to the nodes in a tree

Usage

NodeNumbers(tree, tips = FALSE)

Arguments

tree

A tree of class phylo.

tips

Logical specifying whether to also include the indices of leaves.

Value

NodeNumbers() returns an integer vector corresponding to the indices of nodes within a tree.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)