Parse TNT Tree

Reads a tree from TNT's parenthetical output.

ReadTntTree(filename, relativePath = NULL, keepEnd = 1L, tipLabels = NULL)

TNTText2Tree(treeText)

Arguments

filename
character string specifying path to TNT .tre file.

relativePath
(optional) character string specifying location of the matrix file used to generate the TNT results, relative to the current working directory, for portability. Taxon names will be read from this file if they are not specified by tipLabels.

keepEnd
(optional, default 1) integer specifying how many elements of the file path to conserve when creating relative path (see examples).

tipLabels
(optional) character vector specifying the names of the taxa, in the sequence that they appear in the TNT file. If not specified, taxon names will be loaded from the data file linked in the first line of the .tre file specified in filename.

treeText
Character string describing a tree, in the parenthetical format output by TNT.

Value

ReadTNTTree returns a tree of class phylo, corresponding to the tree in filename.

Details

Supports trees that have been written from TNT in parenthetical notation using tsav*. If taxa have been saved using their names (taxname=), then the tip labels will be read directly from the TNT .tre file. If taxa have been saved using their numbers (taxname-), tip labels will be imported from the linked matrix file that is listed in the first line of the .tre file. Ensure that this file exists in the expected location -- if it does not, use the relativePath argument to override this default, or specify tipLabels to set manually.

Functions

TNTText2Tree: Converts text representation of a tree in TNT to an object of class phylo.

Examples

  # In the examples below, TNT has read a matrix from
  # "c:/TreeTools/input/dataset.nex"
  # The results of an analysis were written to
  # "c:/TreeTools/output/results1.tnt"
  #
  # results1.tnt will contain a hard-coded reference to
  # "c:/TreeTools/input/dataset.nex".

  # On the original machine (but not elsewhere), it would be possible to read
  # this hard-coded reference from results.tnt:
  # ReadTntTree('output/results1.tnt')

  # These datasets are provided with the `TreeTools` package, which will
  # probably not be located at c:/TreeTools on your machine:

  oldWD <- getwd() # Remember the current working directory
  setwd(system.file(package = 'TreeTools'))

  # If taxon names were saved within the file (using `taxname=` in TNT),
  # then our job is easy:
  ReadTntTree('extdata/output/named.tre')

#> 
#> Phylogenetic tree with 5 tips and 4 internal nodes.
#> 
#> Tip labels:
#> [1] "taxon_a" "taxon_b" "taxon_c" "taxon_d" "taxon_e"
#> 
#> Rooted; no branch lengths.

  # But if taxa were compressed to numbers (using `taxname-`), we need to
  # look up the original matrix in order to dereference the tip names.
  #
  # We need to extract the relevant file path from the end of the
  # hard-coded path in the original file.
  #
  # We are interested in the last two elements of
  # c:/TreeTools/input/dataset.nex
  #                2      1
  #
  # '.' means "relative to the current directory"
  ReadTntTree('extdata/output/numbered.tre', './extdata', 2)

#> 2 phylogenetic trees

  # If working in a lower subdirectory
  setwd('./extdata/otherfolder')

  # then it will be necessary to navigate up the directory path with '..':
  ReadTntTree('../output/numbered.tre', '..', 2)

#> 2 phylogenetic trees

  setwd(oldWD) # Restore original working directory