Renumber a tree's nodes and tips

Renumber() numbers the nodes and tips in a tree to conform with the phylo standards.

Usage

Renumber(tree)

Arguments

tree

A tree of class phylo.

Value

Renumber() returns a tree of class phylo, numbered in a Cladewise fashion consistent with the expectations of ape functions.

Details

The ape class phylo is not formally defined, but expects trees' internal representation to conform to certain principles: for example, nodes should be numbered sequentially, with values increasing away from the root.

Renumber() attempts to reformat any tree into a representation that will not cause ape functions to produce unwanted results or to crash R.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

tree <- RandomTree(letters[1:10])
Renumber(tree)
#> 
#> Phylogenetic tree with 10 tips and 8 internal nodes.
#> 
#> Tip labels:
#>   d, g, c, i, f, j, ...
#> 
#> Unrooted; no branch length.