Maximum splits in an n-leaf tree

SplitsInBinaryTree() is a convenience function to calculate the number of splits in a fully-resolved (binary) tree with n leaves.

Usage

SplitsInBinaryTree(tree)

# S3 method for class 'list'
SplitsInBinaryTree(tree)

# S3 method for class 'multiPhylo'
SplitsInBinaryTree(tree)

# S3 method for class 'numeric'
SplitsInBinaryTree(tree)

# S3 method for class '`NULL`'
SplitsInBinaryTree(tree)

# Default S3 method
SplitsInBinaryTree(tree)

# S3 method for class 'Splits'
SplitsInBinaryTree(tree)

# S3 method for class 'phylo'
SplitsInBinaryTree(tree)

Arguments

tree

An object of a supported format that represents a tree or set of trees, from which the number of leaves will be calculated.

Value

SplitsInBinaryTree() returns an integer vector detailing the number of unique non-trivial splits in a binary tree with n leaves.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

tree <- BalancedTree(8)
SplitsInBinaryTree(tree)
#> [1] 5
SplitsInBinaryTree(as.Splits(tree))
#> [1] 5
SplitsInBinaryTree(8)
#> [1] 5
SplitsInBinaryTree(list(tree, tree))
#> [1] 5 5