Write a phylogenetic tree in Newick format

as.Newick() creates a character string representation of a phylogenetic tree, in the Newick format, using R's internal tip numbering. Use RenumberTips() to ensure that the internal numbering follows the order you expect.

Usage

as.Newick(x)

# S3 method for class 'phylo'
as.Newick(x)

# S3 method for class 'list'
as.Newick(x)

# S3 method for class 'multiPhylo'
as.Newick(x)

Arguments

x

Object to convert to Newick format. See Usage section for supported classes.

Value

as.Newick() returns a character string representing tree in Newick format.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

trees <- list(BalancedTree(1:8), PectinateTree(8:1))
trees <- lapply(trees, RenumberTips, 1:8)
as.Newick(trees)
#> [1] "(((0,1),(2,3)),((4,5),(6,7)));" "(((((((0,1),2),3),4),5),6),7);"