Read characters from Nexus file

Parses a Nexus file, reading character states and names.

ReadCharacters(filepath, character_num = NULL, session = NULL)

ReadTntCharacters(filepath, character_num = NULL, session = NULL)

ReadAsPhyDat(filepath)

ReadTntAsPhyDat(filepath)

PhyDat(dataset)

Arguments

filepath
character string specifying location of file

character_num
Index of character(s) to return. NULL, the default, returns all characters.

session
(optionally) a Shiny session with a numericInput named character_num whose maximum should be updated.

dataset
list of taxa and characters, in the format produced by read.nexus.data: a list of sequences each made of a single vector of mode character, and named with the taxon name.

Value

A matrix whose row names correspond to tip labels, and column names correspond to character labels, with the attribute state.labels listing the state labels for each character; or a character string explaining why the character cannot be returned.

Details

Tested with nexus files downloaded from MorphoBank with the "no notes" option, but should also work more generally.

Please report incorrectly parsed files.

Functions

ReadTntCharacters: Read characters from TNT file

ReadAsPhyDat: Read Nexus characters as phyDat object.

ReadTntAsPhyDat: Read TNT characters as phyDat object.

PhyDat: A convenient wrapper for phangorn's phyDat(), which converts a list of morphological characters into a phyDat object. If your morphological characters are in the form of a matrix, perhaps because they have been read using read.table, try MatrixToPhyDat() instead.

References

Maddison DR, Swofford DL, Maddison WP (1997). “Nexus: an extensible file format for systematic information.” Systematic Biology, 46, 590--621. doi: 10.1093/sysbio/46.4.590 .

Examples

fileName <- paste0(system.file(package='TreeTools'),
                   '/extdata/input/dataset.nex')
ReadCharacters(fileName)

#>         [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> taxon_a "0"  "0"  "0"  "0"  "0"  "0"  "0" 
#> taxon_b "0"  "0"  "0"  "0"  "0"  "0"  "0" 
#> taxon_c "1"  "1"  "1"  "1"  "?"  "0"  "0" 
#> taxon_d "1"  "1"  "1"  "?"  "?"  "1"  "1" 
#> taxon_e "1"  "1"  "1"  "1"  "?"  "1"  "1"