Write morphological character matrix to TNT file

Write morphological character matrix to TNT file

WriteTntCharacters(
  dataset,
  filepath = NULL,
  comment = "Dataset written by `TreeTools::WriteTntCharacters()`",
  types = NULL,
  pre = "",
  post = ""
)

# S3 method for phyDat
WriteTntCharacters(
  dataset,
  filepath = NULL,
  comment = "Dataset written by `TreeTools::WriteTntCharacters()`",
  types = NULL,
  pre = "",
  post = ""
)

# S3 method for matrix
WriteTntCharacters(
  dataset,
  filepath = NULL,
  comment = "Dataset written by `TreeTools::WriteTntCharacters()`",
  types = NULL,
  pre = "",
  post = ""
)

Arguments

dataset

Morphological dataset of class phyDat or matrix.

filepath

Path to file; if NULL, returns a character vector.

comment

Optional comment with which to entitle matrix.

types

Optional list specifying where different data types begin. c(num = 1, dna = 10) sets characters 1..9 as numeric, 10..end as DNA.

pre, post

Character vector listing text to print before and after the character matrix. Specify pre = 'piwe=; if the matrix is to be analysed using extended implied weighting (xpiwe=).

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

data('Lobo', package = 'TreeTools')

WriteTntCharacters(Lobo.phy)
#> [1] "\nxread 'Dataset written by `TreeTools::WriteTntCharacters()`'\n115 48\nTubiluchus_Priapulida 00----00000-0100111100-----------------0---0001-1000------0----100-----------0-0--00-00----00--000--00--111--000111\nCricocosmia 00----00000-0100111?00-----------------0---0001-1001100-1001-00100-----------0-0--00-00----00--0???????????????0???\nAysheaia 10----00000-0100?0???0--010-00-0-0-00--10-1000111000------0----1000-0--0100000104000-00---0011-1???????????????0???\nSiberion 10----00000-0100?0???0--010-00-0-0-010-????00011?000------?----1?00-0--0???00??????0-00---?0???????????????????????\nOnychodictyon_ferox 10----00000-0????0???100010-00-0-0-00--1110000111101101010?1-001000-0--011-000101100-00---0011-1???????????????????\nOnychodictyon_gracilis 10----0??00-0????0??????0???0?-?-0-????????000110?0110??1??1-0010?0-0--0???000101?00-00??-1011-0???????????????????\nDiania 10----00000-?????0???0--00--00-0-0-00--0---000110101100-0??0-001000-0--0100000??--00-00---100--????????????????????\nXenusion 10----0?00???????????0--00--00-0-0-00--????00011?101100-0??1-001?00-0--01??????????0-00---100--????????????????????\nPaucipodia 10----00000-?????01??0--00--00-0-0-00--0---000110000------?----0000-0--00--000101000-00---100--????????????????????\nMicrodictyon 10----01000-?????0???0--00--00-0-0-00--0---000111101100-10?1-000000-0--00--000101?00-00---100--0???????????????????\nCardiodictyon 10----0110d??????0???10?0???0?-??0-00--0---000111101100-10?1-000000-0--00--000101?00-00---100--0???????????????????\nHallucigenia_sparsa 10----01100-11001010110000--00-d-0-00--0---0000----111111111000-010-0--00--000101000-0010-000--0???????????????????\nHallucigenia_fortis 10----01100-?????0???10?00--00-d-0-00--0---0001111?111111??10000010-0--00--000101?00-0000-?00--0???????????????????\nHallucigenia_hongmeia 1???????????????????????0???0?-??0-????????00011?1?1111110?10100??0-0--00--000100-00-00??-100--????????????????????\nLuolishania 1100---?100-?????0???100010-00-d-0-00--0---00011110111111??20111010-0--0101100100-00-0021-100--0???????????????????\nCollinsium 1100---1000-?????????0--010-00-d-0-00--0---000110101111110140111010-0--0101000100-00-0031-100--0???????????????????\nCollins_monster_Burgess_Shale 1???????????????????????0???00-d-0-????????00011?10111111????????10-0--0101????????0-0031-?0???????????????????????\nCollins_monster_Emu_Bay 110?????????????????????0???0?-d-0-????????00011?10111111??201???10-0--0101000100-00-0021-?0???????????????????????\nAcinocrinus 1???????????????????????0???00-d-0-????????00011?10111111??511??010-0--0101????????0-0021-?0???????????????????????\nOrstenotubulus 1???????????????????????0???0?-???-????????00011?11111?????101?1??0-0--00--11??????0-00??-?0???????????????????????\nTritonychus 1???????????????????????0???0?-???-????????00011??1????????????1??0-0--00--1101?2000-00????????????????????????01??\nCarbotubulus 1???????????????????????0???0?-???-????????0000----????????????-?10-0--00--00??????0-00??-00?e-????????????????0???\nAntennacanthopodia 10----0?00???????????100010-01-1-0-00--0---0001????0------?----??00-0--00--1?000--00-00---100--????????????????????\nHelenodora 10----0?00??????????????010-0?-??1-00--0---000?????0------?----10?0-0--00--1?01010?0-00??-?????????????????????????\nEuperipatoides_Onychophora 10----00001110--100--100010-01-201-00--0---000110010------1----1000-0--00--110101010-00---100--00112101111101010111\nPlicatoperipatus_Onychophora 10----00001110--100--100010-01-211-00--0---000110010------1----1000-0--00--110101010-00---100--00112101111101010111\nOoperipatellus_Onychophora 10----00001110--100--100010-01-211-00--0---000110010------1----1000-0--00--110101010-00---100--00112101111101010111\nActinarctus_Heterotardigrada 1100---00110010010100100011000-0-0-10--0---0000----1100-100??---000-0--00--001103100-00---0011-11100110111011?-100-\nHalobiotus_Eutardigrada 10----00000-010010100100011000-0-0-20--0---0000----10-----0??---000-0--00--001111100-00---0011-111001101110111-100-\nSiberian_Orsten_tardigrade 10----00000-0100??????b?0???0?-???-????????0000----???--?????---?00-0--00--000111100-00---001a-????????????????????\nMegadictyon 10----0?00???1???0???0--010-00-0-0-00--10-100011?0?0------?----0100-0--011-000????00-00---?0???????????????????????\nJianshanopodia 10----00000-01??101100--010-00-0-0-0???10-?00011?100------?----0100-0--011-000????00-00---00101????????????????????\nHadranax 1???????????????????????0???0?-???-????????00011?101100-0-?3--01?00-0--00--10?????00-00---?0???????????????????????\nKerygmachela 10----00000-0100?0??????010-00-0-0-010-10-1000110101100-0-?3-?01101001000--00000--00000---00100????????????????0???\nPambdelurion 10----0000100100?0???0--010-00-0-0-110-10-?0001????????????--?-?101001000--00000--00000---?????????????????????110-\nOpabinia 10----0000210????0???111010-00-0-0-111010-11000----0------------101001100--00000--00000---?1101????????????????????\nAnomalocaris_canadensis 11010000?011011110???111011110-0-0-110-11010000----0------------101000100--00000--01101---01101????????????????1???\nPeytoia_nathorsti 11011000?0110111?0???111011110-0-0-110-10-10000----0------------101001100--00000--01101---000--????????????????????\nHurdia_victoria 11011010?011011110110111011110-0-0-1???10-1?000----0------------?01001100--00000--01001---00101????????????????????\nAegirocassis_benmoulae 11011010?0?10???????????011110-0-0-1???10-10000----0------------?01001100--00000--0110?---000--????????????????????\nLyrarapax_unguispinus 11010000?01101???????111011110-0-0-110-10-10000----0------------001000100--00000--0?111---01101??10011?????????1???\nSchinderhannes 1??????0?01101???????1110111??-0-0-110-110???????????????????????????????????????????1?---00101????????????????????\nFuxianhuia 110101-0--2100--?0???1111110-10---11111----110------------------0-111-00-----000--00-00--1-0101??12111????????1100-\nChengjiangocaris 110101-0--2100--?0???1111110-10---11111----110------------------0-111-00-----000--00-00--1-0101?1?????0?1??????????\nLeanchoilia 1110---0--2100--?0???1111110-11---01111----11?------------0-----1-111-01-----010c?00-00--0-00--????????????????????\nAlalcomenaeus 1110---0--2100--?0???11111??-11---01111----11?------------0-----1-111-01-----0100-00-00--0-00--?112111000--????????\nMisszhouia_longicaudata 111101-0--2100--?0???1111110-10---01111----111------------0-----1-121-01-----0100-00-00--0-00--????????????????1???\nKuamaia_lata 111101-0--2100--?0???1111110-10---01111----111------------0-----1-121-01-----0102100-00--0-00--????????????????1???\n;\n"

# Read with extended implied weighting
WriteTntCharacters(Lobo.phy, pre = 'piwe=10;', post = 'xpiwe=;')
#> [1] "piwe=10;\nxread 'Dataset written by `TreeTools::WriteTntCharacters()`'\n115 48\nTubiluchus_Priapulida 00----00000-0100111100-----------------0---0001-1000------0----100-----------0-0--00-00----00--000--00--111--000111\nCricocosmia 00----00000-0100111?00-----------------0---0001-1001100-1001-00100-----------0-0--00-00----00--0???????????????0???\nAysheaia 10----00000-0100?0???0--010-00-0-0-00--10-1000111000------0----1000-0--0100000104000-00---0011-1???????????????0???\nSiberion 10----00000-0100?0???0--010-00-0-0-010-????00011?000------?----1?00-0--0???00??????0-00---?0???????????????????????\nOnychodictyon_ferox 10----00000-0????0???100010-00-0-0-00--1110000111101101010?1-001000-0--011-000101100-00---0011-1???????????????????\nOnychodictyon_gracilis 10----0??00-0????0??????0???0?-?-0-????????000110?0110??1??1-0010?0-0--0???000101?00-00??-1011-0???????????????????\nDiania 10----00000-?????0???0--00--00-0-0-00--0---000110101100-0??0-001000-0--0100000??--00-00---100--????????????????????\nXenusion 10----0?00???????????0--00--00-0-0-00--????00011?101100-0??1-001?00-0--01??????????0-00---100--????????????????????\nPaucipodia 10----00000-?????01??0--00--00-0-0-00--0---000110000------?----0000-0--00--000101000-00---100--????????????????????\nMicrodictyon 10----01000-?????0???0--00--00-0-0-00--0---000111101100-10?1-000000-0--00--000101?00-00---100--0???????????????????\nCardiodictyon 10----0110d??????0???10?0???0?-??0-00--0---000111101100-10?1-000000-0--00--000101?00-00---100--0???????????????????\nHallucigenia_sparsa 10----01100-11001010110000--00-d-0-00--0---0000----111111111000-010-0--00--000101000-0010-000--0???????????????????\nHallucigenia_fortis 10----01100-?????0???10?00--00-d-0-00--0---0001111?111111??10000010-0--00--000101?00-0000-?00--0???????????????????\nHallucigenia_hongmeia 1???????????????????????0???0?-??0-????????00011?1?1111110?10100??0-0--00--000100-00-00??-100--????????????????????\nLuolishania 1100---?100-?????0???100010-00-d-0-00--0---00011110111111??20111010-0--0101100100-00-0021-100--0???????????????????\nCollinsium 1100---1000-?????????0--010-00-d-0-00--0---000110101111110140111010-0--0101000100-00-0031-100--0???????????????????\nCollins_monster_Burgess_Shale 1???????????????????????0???00-d-0-????????00011?10111111????????10-0--0101????????0-0031-?0???????????????????????\nCollins_monster_Emu_Bay 110?????????????????????0???0?-d-0-????????00011?10111111??201???10-0--0101000100-00-0021-?0???????????????????????\nAcinocrinus 1???????????????????????0???00-d-0-????????00011?10111111??511??010-0--0101????????0-0021-?0???????????????????????\nOrstenotubulus 1???????????????????????0???0?-???-????????00011?11111?????101?1??0-0--00--11??????0-00??-?0???????????????????????\nTritonychus 1???????????????????????0???0?-???-????????00011??1????????????1??0-0--00--1101?2000-00????????????????????????01??\nCarbotubulus 1???????????????????????0???0?-???-????????0000----????????????-?10-0--00--00??????0-00??-00?e-????????????????0???\nAntennacanthopodia 10----0?00???????????100010-01-1-0-00--0---0001????0------?----??00-0--00--1?000--00-00---100--????????????????????\nHelenodora 10----0?00??????????????010-0?-??1-00--0---000?????0------?----10?0-0--00--1?01010?0-00??-?????????????????????????\nEuperipatoides_Onychophora 10----00001110--100--100010-01-201-00--0---000110010------1----1000-0--00--110101010-00---100--00112101111101010111\nPlicatoperipatus_Onychophora 10----00001110--100--100010-01-211-00--0---000110010------1----1000-0--00--110101010-00---100--00112101111101010111\nOoperipatellus_Onychophora 10----00001110--100--100010-01-211-00--0---000110010------1----1000-0--00--110101010-00---100--00112101111101010111\nActinarctus_Heterotardigrada 1100---00110010010100100011000-0-0-10--0---0000----1100-100??---000-0--00--001103100-00---0011-11100110111011?-100-\nHalobiotus_Eutardigrada 10----00000-010010100100011000-0-0-20--0---0000----10-----0??---000-0--00--001111100-00---0011-111001101110111-100-\nSiberian_Orsten_tardigrade 10----00000-0100??????b?0???0?-???-????????0000----???--?????---?00-0--00--000111100-00---001a-????????????????????\nMegadictyon 10----0?00???1???0???0--010-00-0-0-00--10-100011?0?0------?----0100-0--011-000????00-00---?0???????????????????????\nJianshanopodia 10----00000-01??101100--010-00-0-0-0???10-?00011?100------?----0100-0--011-000????00-00---00101????????????????????\nHadranax 1???????????????????????0???0?-???-????????00011?101100-0-?3--01?00-0--00--10?????00-00---?0???????????????????????\nKerygmachela 10----00000-0100?0??????010-00-0-0-010-10-1000110101100-0-?3-?01101001000--00000--00000---00100????????????????0???\nPambdelurion 10----0000100100?0???0--010-00-0-0-110-10-?0001????????????--?-?101001000--00000--00000---?????????????????????110-\nOpabinia 10----0000210????0???111010-00-0-0-111010-11000----0------------101001100--00000--00000---?1101????????????????????\nAnomalocaris_canadensis 11010000?011011110???111011110-0-0-110-11010000----0------------101000100--00000--01101---01101????????????????1???\nPeytoia_nathorsti 11011000?0110111?0???111011110-0-0-110-10-10000----0------------101001100--00000--01101---000--????????????????????\nHurdia_victoria 11011010?011011110110111011110-0-0-1???10-1?000----0------------?01001100--00000--01001---00101????????????????????\nAegirocassis_benmoulae 11011010?0?10???????????011110-0-0-1???10-10000----0------------?01001100--00000--0110?---000--????????????????????\nLyrarapax_unguispinus 11010000?01101???????111011110-0-0-110-10-10000----0------------001000100--00000--0?111---01101??10011?????????1???\nSchinderhannes 1??????0?01101???????1110111??-0-0-110-110???????????????????????????????????????????1?---00101????????????????????\nFuxianhuia 110101-0--2100--?0???1111110-10---11111----110------------------0-111-00-----000--00-00--1-0101??12111????????1100-\nChengjiangocaris 110101-0--2100--?0???1111110-10---11111----110------------------0-111-00-----000--00-00--1-0101?1?????0?1??????????\nLeanchoilia 1110---0--2100--?0???1111110-11---01111----11?------------0-----1-111-01-----010c?00-00--0-00--????????????????????\nAlalcomenaeus 1110---0--2100--?0???11111??-11---01111----11?------------0-----1-111-01-----0100-00-00--0-00--?112111000--????????\nMisszhouia_longicaudata 111101-0--2100--?0???1111110-10---01111----111------------0-----1-121-01-----0100-00-00--0-00--????????????????1???\nKuamaia_lata 111101-0--2100--?0???1111110-10---01111----111------------0-----1-121-01-----0102100-00--0-00--????????????????1???\n;\nxpiwe=;"

# Write to a file with:
# WriteTntCharacters(Lobo.phy, 'example_file.tnt')