Write morphological character matrix to TNT file

Write morphological character matrix to TNT file

Usage

WriteTntCharacters(
  dataset,
  filepath = NULL,
  comment = "Dataset written by `TreeTools::WriteTntCharacters()`",
  types = NULL,
  pre = "",
  post = ""
)

# S3 method for class 'phyDat'
WriteTntCharacters(
  dataset,
  filepath = NULL,
  comment = "Dataset written by `TreeTools::WriteTntCharacters()`",
  types = NULL,
  pre = "",
  post = ""
)

# S3 method for class 'matrix'
WriteTntCharacters(
  dataset,
  filepath = NULL,
  comment = "Dataset written by `TreeTools::WriteTntCharacters()`",
  types = NULL,
  pre = "",
  post = ""
)

Arguments

dataset

Morphological dataset of class phyDat or matrix.

filepath

Path to file; if NULL, returns a character vector.

comment

Optional comment with which to entitle matrix.

types

Optional list specifying where different data types begin. c(num = 1, dna = 10) sets characters 1..9 as numeric, 10..end as DNA.

pre, post

Character vector listing text to print before and after the character matrix. Specify pre = 'piwe=; if the matrix is to be analysed using extended implied weighting (xpiwe=).

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

data("Lobo", package = "TreeTools")

WriteTntCharacters(Lobo.phy)
#> [1] "\nxread 'Dataset written by `TreeTools::WriteTntCharacters()`'\n115 48\nTubiluchus_Priapulida 0 0 - - - - 0 0 0 0 0 - 0 1 0 0 1 1 1 1 0 0 - - - - - - - - - - - - - - - - - 0 - - - 0 0 0 1 - 1 0 0 0 - - - - - - 0 - - - - 1 0 0 - - - - - - - - - - - 0 - 0 - - 0 0 - 0 0 - - - - 0 0 - - 0 0 0 - - 0 0 - - 1 1 1 - - 0 0 0 1 1 1\nCricocosmia 0 0 - - - - 0 0 0 0 0 - 0 1 0 0 1 1 1 ? 0 0 - - - - - - - - - - - - - - - - - 0 - - - 0 0 0 1 - 1 0 0 1 1 0 0 - 1 0 0 1 - 0 0 1 0 0 - - - - - - - - - - - 0 - 0 - - 0 0 - 0 0 - - - - 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nAysheaia 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 0 - - 1 0 - 1 0 0 0 1 1 1 0 0 0 - - - - - - 0 - - - - 1 0 0 0 - 0 - - 0 1 0 0 0 0 0 1 0 4 0 0 0 - 0 0 - - - 0 0 1 1 - 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nSiberion 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 1 0 - ? ? ? ? 0 0 0 1 1 ? 0 0 0 - - - - - - ? - - - - 1 ? 0 0 - 0 - - 0 ? ? ? 0 0 ? ? ? ? ? ? 0 - 0 0 - - - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nOnychodictyon_ferox 1 0 - - - - 0 0 0 0 0 - 0 ? ? ? ? 0 ? ? ? 1 0 0 0 1 0 - 0 0 - 0 - 0 - 0 0 - - 1 1 1 0 0 0 0 1 1 1 1 0 1 1 0 1 0 1 0 ? 1 - 0 0 1 0 0 0 - 0 - - 0 1 1 - 0 0 0 1 0 1 1 0 0 - 0 0 - - - 0 0 1 1 - 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nOnychodictyon_gracilis 1 0 - - - - 0 ? ? 0 0 - 0 ? ? ? ? 0 ? ? ? ? ? ? 0 ? ? ? 0 ? - ? - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 0 ? 0 1 1 0 ? ? 1 ? ? 1 - 0 0 1 0 ? 0 - 0 - - 0 ? ? ? 0 0 0 1 0 1 ? 0 0 - 0 0 ? ? - 1 0 1 1 - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nDiania 1 0 - - - - 0 0 0 0 0 - ? ? ? ? ? 0 ? ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - 0 - - - 0 0 0 1 1 0 1 0 1 1 0 0 - 0 ? ? 0 - 0 0 1 0 0 0 - 0 - - 0 1 0 0 0 0 0 ? ? - - 0 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nXenusion 1 0 - - - - 0 ? 0 0 ? ? ? ? ? ? ? ? ? ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 0 0 - 0 ? ? 1 - 0 0 1 ? 0 0 - 0 - - 0 1 ? ? ? ? ? ? ? ? ? ? 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nPaucipodia 1 0 - - - - 0 0 0 0 0 - ? ? ? ? ? 0 1 ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 0 0 - - - - - - ? - - - - 0 0 0 0 - 0 - - 0 0 - - 0 0 0 1 0 1 0 0 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nMicrodictyon 1 0 - - - - 0 1 0 0 0 - ? ? ? ? ? 0 ? ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 0 1 1 0 0 - 1 0 ? 1 - 0 0 0 0 0 0 - 0 - - 0 0 - - 0 0 0 1 0 1 ? 0 0 - 0 0 - - - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCardiodictyon 1 0 - - - - 0 1 1 0 [01] ? ? ? ? ? ? 0 ? ? ? 1 0 ? 0 ? ? ? 0 ? - ? ? 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 0 1 1 0 0 - 1 0 ? 1 - 0 0 0 0 0 0 - 0 - - 0 0 - - 0 0 0 1 0 1 ? 0 0 - 0 0 - - - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHallucigenia_sparsa 1 0 - - - - 0 1 1 0 0 - 1 1 0 0 1 0 1 0 1 1 0 0 0 0 - - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 0 - - - - 1 1 1 1 1 1 1 1 1 0 0 0 - 0 1 0 - 0 - - 0 0 - - 0 0 0 1 0 1 0 0 0 - 0 0 1 0 - 0 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHallucigenia_fortis 1 0 - - - - 0 1 1 0 0 - ? ? ? ? ? 0 ? ? ? 1 0 ? 0 0 - - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 ? 1 1 1 1 1 1 ? ? 1 0 0 0 0 0 1 0 - 0 - - 0 0 - - 0 0 0 1 0 1 ? 0 0 - 0 0 0 0 - ? 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHallucigenia_hongmeia 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 ? 1 1 1 1 1 1 0 ? 1 0 1 0 0 ? ? 0 - 0 - - 0 0 - - 0 0 0 1 0 0 - 0 0 - 0 0 ? ? - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nLuolishania 1 1 0 0 - - - ? 1 0 0 - ? ? ? ? ? 0 ? ? ? 1 0 0 0 1 0 - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 0 1 1 1 1 1 1 ? ? 2 0 1 1 1 0 1 0 - 0 - - 0 1 0 1 1 0 0 1 0 0 - 0 0 - 0 0 2 1 - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCollinsium 1 1 0 0 - - - 1 0 0 0 - ? ? ? ? ? ? ? ? ? 0 - - 0 1 0 - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 1 1 0 1 0 1 1 1 1 1 1 0 1 4 0 1 1 1 0 1 0 - 0 - - 0 1 0 1 0 0 0 1 0 0 - 0 0 - 0 0 3 1 - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCollins_monster_Burgess_Shale 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 0 - [01] - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 1 1 1 1 ? ? ? ? ? ? ? ? 1 0 - 0 - - 0 1 0 1 ? ? ? ? ? ? ? ? 0 - 0 0 3 1 - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCollins_monster_Emu_Bay 1 1 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - [01] - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 1 1 1 1 ? ? 2 0 1 ? ? ? 1 0 - 0 - - 0 1 0 1 0 0 0 1 0 0 - 0 0 - 0 0 2 1 - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAcinocrinus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 0 - [01] - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 1 1 1 1 ? ? 5 1 1 ? ? 0 1 0 - 0 - - 0 1 0 1 ? ? ? ? ? ? ? ? 0 - 0 0 2 1 - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nOrstenotubulus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 1 1 1 1 ? ? ? ? ? 1 0 1 ? 1 ? ? 0 - 0 - - 0 0 - - 1 1 ? ? ? ? ? ? 0 - 0 0 ? ? - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nTritonychus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? ? 1 ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? 0 - 0 - - 0 0 - - 1 1 0 1 ? 2 0 0 0 - 0 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 1 ? ?\nCarbotubulus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 0 - - - - ? ? ? ? ? ? ? ? ? ? ? ? - ? 1 0 - 0 - - 0 0 - - 0 0 ? ? ? ? ? ? 0 - 0 0 ? ? - 0 0 ? [1-] - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nAntennacanthopodia 1 0 - - - - 0 ? 0 0 ? ? ? ? ? ? ? ? ? ? ? 1 0 0 0 1 0 - 0 1 - 1 - 0 - 0 0 - - 0 - - - 0 0 0 1 ? ? ? ? 0 - - - - - - ? - - - - ? ? 0 0 - 0 - - 0 0 - - 1 ? 0 0 0 - - 0 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHelenodora 1 0 - - - - 0 ? 0 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 1 0 - 0 ? - ? ? 1 - 0 0 - - 0 - - - 0 0 0 ? ? ? ? ? 0 - - - - - - ? - - - - 1 0 ? 0 - 0 - - 0 0 - - 1 ? 0 1 0 1 0 ? 0 - 0 0 ? ? - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nEuperipatoides_Onychophora 1 0 - - - - 0 0 0 0 1 1 1 0 - - 1 0 0 - - 1 0 0 0 1 0 - 0 1 - 2 0 1 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 1 0 - - - - - - 1 - - - - 1 0 0 0 - 0 - - 0 0 - - 1 1 0 1 0 1 0 1 0 - 0 0 - - - 1 0 0 - - 0 0 1 1 2 1 0 1 1 1 1 1 0 1 0 1 0 1 1 1\nPlicatoperipatus_Onychophora 1 0 - - - - 0 0 0 0 1 1 1 0 - - 1 0 0 - - 1 0 0 0 1 0 - 0 1 - 2 1 1 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 1 0 - - - - - - 1 - - - - 1 0 0 0 - 0 - - 0 0 - - 1 1 0 1 0 1 0 1 0 - 0 0 - - - 1 0 0 - - 0 0 1 1 2 1 0 1 1 1 1 1 0 1 0 1 0 1 1 1\nOoperipatellus_Onychophora 1 0 - - - - 0 0 0 0 1 1 1 0 - - 1 0 0 - - 1 0 0 0 1 0 - 0 1 - 2 1 1 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 1 0 - - - - - - 1 - - - - 1 0 0 0 - 0 - - 0 0 - - 1 1 0 1 0 1 0 1 0 - 0 0 - - - 1 0 0 - - 0 0 1 1 2 1 0 1 1 1 1 1 0 1 0 1 0 1 1 1\nActinarctus_Heterotardigrada 1 1 0 0 - - - 0 0 1 1 0 0 1 0 0 1 0 1 0 0 1 0 0 0 1 1 0 0 0 - 0 - 0 - 1 0 - - 0 - - - 0 0 0 0 - - - - 1 1 0 0 - 1 0 0 ? ? - - - 0 0 0 - 0 - - 0 0 - - 0 0 1 1 0 3 1 0 0 - 0 0 - - - 0 0 1 1 - 1 1 1 0 0 1 1 0 1 1 1 0 1 1 ? - 1 0 0 -\nHalobiotus_Eutardigrada 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 1 0 1 0 0 1 0 0 0 1 1 0 0 0 - 0 - 0 - 2 0 - - 0 - - - 0 0 0 0 - - - - 1 0 - - - - - 0 ? ? - - - 0 0 0 - 0 - - 0 0 - - 0 0 1 1 1 1 1 0 0 - 0 0 - - - 0 0 1 1 - 1 1 1 0 0 1 1 0 1 1 1 0 1 1 1 - 1 0 0 -\nSiberian_Orsten_tardigrade 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? ? ? ? ? ? [0-] ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 0 - - - - ? ? ? - - ? ? ? ? ? - - - ? 0 0 - 0 - - 0 0 - - 0 0 0 1 1 1 1 0 0 - 0 0 - - - 0 0 1 [12] - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nMegadictyon 1 0 - - - - 0 ? 0 0 ? ? ? 1 ? ? ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 0 - - 1 0 - 1 0 0 0 1 1 ? 0 ? 0 - - - - - - ? - - - - 0 1 0 0 - 0 - - 0 1 1 - 0 0 0 ? ? ? ? 0 0 - 0 0 - - - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nJianshanopodia 1 0 - - - - 0 0 0 0 0 - 0 1 ? ? 1 0 1 1 0 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 ? ? ? 1 0 - ? 0 0 0 1 1 ? 1 0 0 - - - - - - ? - - - - 0 1 0 0 - 0 - - 0 1 1 - 0 0 0 ? ? ? ? 0 0 - 0 0 - - - 0 0 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHadranax 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 0 0 - 0 - ? 3 - - 0 1 ? 0 0 - 0 - - 0 0 - - 1 0 ? ? ? ? ? 0 0 - 0 0 - - - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nKerygmachela 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? 0 ? ? ? ? ? ? 0 1 0 - 0 0 - 0 - 0 - 0 1 0 - 1 0 - 1 0 0 0 1 1 0 1 0 1 1 0 0 - 0 - ? 3 - ? 0 1 1 0 1 0 0 1 0 0 0 - - 0 0 0 0 0 - - 0 0 0 0 0 - - - 0 0 1 0 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nPambdelurion 1 0 - - - - 0 0 0 0 1 0 0 1 0 0 ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 1 1 0 - 1 0 - ? 0 0 0 1 ? ? ? ? ? ? ? ? ? ? ? ? - - ? - ? 1 0 1 0 0 1 0 0 0 - - 0 0 0 0 0 - - 0 0 0 0 0 - - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 1 0 -\nOpabinia 1 0 - - - - 0 0 0 0 2 1 0 ? ? ? ? 0 ? ? ? 1 1 1 0 1 0 - 0 0 - 0 - 0 - 1 1 1 0 1 0 - 1 1 0 0 0 - - - - 0 - - - - - - - - - - - - 1 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 0 0 0 0 - - - ? 1 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAnomalocaris_canadensis 1 1 0 1 0 0 0 0 ? 0 1 1 0 1 1 1 1 0 ? ? ? 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 1 0 - 1 1 0 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - 1 0 1 0 0 0 1 0 0 - - 0 0 0 0 0 - - 0 1 1 0 1 - - - 0 1 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? ?\nPeytoia_nathorsti 1 1 0 1 1 0 0 0 ? 0 1 1 0 1 1 1 ? 0 ? ? ? 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 1 0 - 1 0 - 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - 1 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 1 1 0 1 - - - 0 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHurdia_victoria 1 1 0 1 1 0 1 0 ? 0 1 1 0 1 1 1 1 0 1 1 0 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 ? ? ? 1 0 - 1 ? 0 0 0 - - - - 0 - - - - - - - - - - - - ? 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 1 0 0 1 - - - 0 0 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAegirocassis_benmoulae 1 1 0 1 1 0 1 0 ? 0 ? 1 0 ? ? ? ? ? ? ? ? ? ? ? 0 1 1 1 1 0 - 0 - 0 - 1 ? ? ? 1 0 - 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - ? 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 1 1 0 ? - - - 0 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nLyrarapax_unguispinus 1 1 0 1 0 0 0 0 ? 0 1 1 0 1 ? ? ? ? ? ? ? 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 1 0 - 1 0 - 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - 0 0 1 0 0 0 1 0 0 - - 0 0 0 0 0 - - 0 ? 1 1 1 - - - 0 1 1 0 1 ? ? 1 0 0 1 1 ? ? ? ? ? ? ? ? ? 1 ? ? ?\nSchinderhannes 1 ? ? ? ? ? ? 0 ? 0 1 1 0 1 ? ? ? ? ? ? ? 1 1 1 0 1 1 1 ? ? - 0 - 0 - 1 1 0 - 1 1 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? - - - 0 0 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nFuxianhuia 1 1 0 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 1 1 1 1 1 - - - - 1 1 0 - - - - - - - - - - - - - - - - - - 0 - 1 1 1 - 0 0 - - - - - 0 0 0 - - 0 0 - 0 0 - - 1 - 0 1 0 1 ? ? 1 2 1 1 1 ? ? ? ? ? ? ? ? 1 1 0 0 -\nChengjiangocaris 1 1 0 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 1 1 1 1 1 - - - - 1 1 0 - - - - - - - - - - - - - - - - - - 0 - 1 1 1 - 0 0 - - - - - 0 0 0 - - 0 0 - 0 0 - - 1 - 0 1 0 1 ? 1 ? ? ? ? ? 0 ? 1 ? ? ? ? ? ? ? ? ? ?\nLeanchoilia 1 1 1 0 - - - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 1 - - - 0 1 1 1 1 - - - - 1 1 ? - - - - - - - - - - - - 0 - - - - - 1 - 1 1 1 - 0 1 - - - - - 0 1 0 [02] ? 0 0 - 0 0 - - 0 - 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAlalcomenaeus 1 1 1 0 - - - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 ? ? - 1 1 - - - 0 1 1 1 1 - - - - 1 1 ? - - - - - - - - - - - - 0 - - - - - 1 - 1 1 1 - 0 1 - - - - - 0 1 0 0 - 0 0 - 0 0 - - 0 - 0 0 - - ? 1 1 2 1 1 1 0 0 0 - - ? ? ? ? ? ? ? ?\nMisszhouia_longicaudata 1 1 1 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 0 1 1 1 1 - - - - 1 1 1 - - - - - - - - - - - - 0 - - - - - 1 - 1 2 1 - 0 1 - - - - - 0 1 0 0 - 0 0 - 0 0 - - 0 - 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? ?\nKuamaia_lata 1 1 1 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 0 1 1 1 1 - - - - 1 1 1 - - - - - - - - - - - - 0 - - - - - 1 - 1 2 1 - 0 1 - - - - - 0 1 0 2 1 0 0 - 0 0 - - 0 - 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? ?\n;\n"

# Read with extended implied weighting
WriteTntCharacters(Lobo.phy, pre = "piwe=10;", post = "xpiwe=;")
#> [1] "piwe=10;\nxread 'Dataset written by `TreeTools::WriteTntCharacters()`'\n115 48\nTubiluchus_Priapulida 0 0 - - - - 0 0 0 0 0 - 0 1 0 0 1 1 1 1 0 0 - - - - - - - - - - - - - - - - - 0 - - - 0 0 0 1 - 1 0 0 0 - - - - - - 0 - - - - 1 0 0 - - - - - - - - - - - 0 - 0 - - 0 0 - 0 0 - - - - 0 0 - - 0 0 0 - - 0 0 - - 1 1 1 - - 0 0 0 1 1 1\nCricocosmia 0 0 - - - - 0 0 0 0 0 - 0 1 0 0 1 1 1 ? 0 0 - - - - - - - - - - - - - - - - - 0 - - - 0 0 0 1 - 1 0 0 1 1 0 0 - 1 0 0 1 - 0 0 1 0 0 - - - - - - - - - - - 0 - 0 - - 0 0 - 0 0 - - - - 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nAysheaia 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 0 - - 1 0 - 1 0 0 0 1 1 1 0 0 0 - - - - - - 0 - - - - 1 0 0 0 - 0 - - 0 1 0 0 0 0 0 1 0 4 0 0 0 - 0 0 - - - 0 0 1 1 - 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nSiberion 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 1 0 - ? ? ? ? 0 0 0 1 1 ? 0 0 0 - - - - - - ? - - - - 1 ? 0 0 - 0 - - 0 ? ? ? 0 0 ? ? ? ? ? ? 0 - 0 0 - - - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nOnychodictyon_ferox 1 0 - - - - 0 0 0 0 0 - 0 ? ? ? ? 0 ? ? ? 1 0 0 0 1 0 - 0 0 - 0 - 0 - 0 0 - - 1 1 1 0 0 0 0 1 1 1 1 0 1 1 0 1 0 1 0 ? 1 - 0 0 1 0 0 0 - 0 - - 0 1 1 - 0 0 0 1 0 1 1 0 0 - 0 0 - - - 0 0 1 1 - 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nOnychodictyon_gracilis 1 0 - - - - 0 ? ? 0 0 - 0 ? ? ? ? 0 ? ? ? ? ? ? 0 ? ? ? 0 ? - ? - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 0 ? 0 1 1 0 ? ? 1 ? ? 1 - 0 0 1 0 ? 0 - 0 - - 0 ? ? ? 0 0 0 1 0 1 ? 0 0 - 0 0 ? ? - 1 0 1 1 - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nDiania 1 0 - - - - 0 0 0 0 0 - ? ? ? ? ? 0 ? ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - 0 - - - 0 0 0 1 1 0 1 0 1 1 0 0 - 0 ? ? 0 - 0 0 1 0 0 0 - 0 - - 0 1 0 0 0 0 0 ? ? - - 0 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nXenusion 1 0 - - - - 0 ? 0 0 ? ? ? ? ? ? ? ? ? ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 0 0 - 0 ? ? 1 - 0 0 1 ? 0 0 - 0 - - 0 1 ? ? ? ? ? ? ? ? ? ? 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nPaucipodia 1 0 - - - - 0 0 0 0 0 - ? ? ? ? ? 0 1 ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 0 0 - - - - - - ? - - - - 0 0 0 0 - 0 - - 0 0 - - 0 0 0 1 0 1 0 0 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nMicrodictyon 1 0 - - - - 0 1 0 0 0 - ? ? ? ? ? 0 ? ? ? 0 - - 0 0 - - 0 0 - 0 - 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 0 1 1 0 0 - 1 0 ? 1 - 0 0 0 0 0 0 - 0 - - 0 0 - - 0 0 0 1 0 1 ? 0 0 - 0 0 - - - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCardiodictyon 1 0 - - - - 0 1 1 0 [01] ? ? ? ? ? ? 0 ? ? ? 1 0 ? 0 ? ? ? 0 ? - ? ? 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 0 1 1 0 0 - 1 0 ? 1 - 0 0 0 0 0 0 - 0 - - 0 0 - - 0 0 0 1 0 1 ? 0 0 - 0 0 - - - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHallucigenia_sparsa 1 0 - - - - 0 1 1 0 0 - 1 1 0 0 1 0 1 0 1 1 0 0 0 0 - - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 0 - - - - 1 1 1 1 1 1 1 1 1 0 0 0 - 0 1 0 - 0 - - 0 0 - - 0 0 0 1 0 1 0 0 0 - 0 0 1 0 - 0 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHallucigenia_fortis 1 0 - - - - 0 1 1 0 0 - ? ? ? ? ? 0 ? ? ? 1 0 ? 0 0 - - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 ? 1 1 1 1 1 1 ? ? 1 0 0 0 0 0 1 0 - 0 - - 0 0 - - 0 0 0 1 0 1 ? 0 0 - 0 0 0 0 - ? 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHallucigenia_hongmeia 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 ? 1 1 1 1 1 1 0 ? 1 0 1 0 0 ? ? 0 - 0 - - 0 0 - - 0 0 0 1 0 0 - 0 0 - 0 0 ? ? - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nLuolishania 1 1 0 0 - - - ? 1 0 0 - ? ? ? ? ? 0 ? ? ? 1 0 0 0 1 0 - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 1 1 1 1 0 1 1 1 1 1 1 ? ? 2 0 1 1 1 0 1 0 - 0 - - 0 1 0 1 1 0 0 1 0 0 - 0 0 - 0 0 2 1 - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCollinsium 1 1 0 0 - - - 1 0 0 0 - ? ? ? ? ? ? ? ? ? 0 - - 0 1 0 - 0 0 - [01] - 0 - 0 0 - - 0 - - - 0 0 0 1 1 0 1 0 1 1 1 1 1 1 0 1 4 0 1 1 1 0 1 0 - 0 - - 0 1 0 1 0 0 0 1 0 0 - 0 0 - 0 0 3 1 - 1 0 0 - - 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCollins_monster_Burgess_Shale 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 0 - [01] - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 1 1 1 1 ? ? ? ? ? ? ? ? 1 0 - 0 - - 0 1 0 1 ? ? ? ? ? ? ? ? 0 - 0 0 3 1 - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nCollins_monster_Emu_Bay 1 1 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - [01] - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 1 1 1 1 ? ? 2 0 1 ? ? ? 1 0 - 0 - - 0 1 0 1 0 0 0 1 0 0 - 0 0 - 0 0 2 1 - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAcinocrinus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 0 - [01] - 0 - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 1 1 1 1 ? ? 5 1 1 ? ? 0 1 0 - 0 - - 0 1 0 1 ? ? ? ? ? ? ? ? 0 - 0 0 2 1 - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nOrstenotubulus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 1 1 1 1 ? ? ? ? ? 1 0 1 ? 1 ? ? 0 - 0 - - 0 0 - - 1 1 ? ? ? ? ? ? 0 - 0 0 ? ? - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nTritonychus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? ? 1 ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? 0 - 0 - - 0 0 - - 1 1 0 1 ? 2 0 0 0 - 0 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 1 ? ?\nCarbotubulus 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 0 - - - - ? ? ? ? ? ? ? ? ? ? ? ? - ? 1 0 - 0 - - 0 0 - - 0 0 ? ? ? ? ? ? 0 - 0 0 ? ? - 0 0 ? [1-] - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nAntennacanthopodia 1 0 - - - - 0 ? 0 0 ? ? ? ? ? ? ? ? ? ? ? 1 0 0 0 1 0 - 0 1 - 1 - 0 - 0 0 - - 0 - - - 0 0 0 1 ? ? ? ? 0 - - - - - - ? - - - - ? ? 0 0 - 0 - - 0 0 - - 1 ? 0 0 0 - - 0 0 - 0 0 - - - 1 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHelenodora 1 0 - - - - 0 ? 0 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 1 0 - 0 ? - ? ? 1 - 0 0 - - 0 - - - 0 0 0 ? ? ? ? ? 0 - - - - - - ? - - - - 1 0 ? 0 - 0 - - 0 0 - - 1 ? 0 1 0 1 0 ? 0 - 0 0 ? ? - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nEuperipatoides_Onychophora 1 0 - - - - 0 0 0 0 1 1 1 0 - - 1 0 0 - - 1 0 0 0 1 0 - 0 1 - 2 0 1 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 1 0 - - - - - - 1 - - - - 1 0 0 0 - 0 - - 0 0 - - 1 1 0 1 0 1 0 1 0 - 0 0 - - - 1 0 0 - - 0 0 1 1 2 1 0 1 1 1 1 1 0 1 0 1 0 1 1 1\nPlicatoperipatus_Onychophora 1 0 - - - - 0 0 0 0 1 1 1 0 - - 1 0 0 - - 1 0 0 0 1 0 - 0 1 - 2 1 1 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 1 0 - - - - - - 1 - - - - 1 0 0 0 - 0 - - 0 0 - - 1 1 0 1 0 1 0 1 0 - 0 0 - - - 1 0 0 - - 0 0 1 1 2 1 0 1 1 1 1 1 0 1 0 1 0 1 1 1\nOoperipatellus_Onychophora 1 0 - - - - 0 0 0 0 1 1 1 0 - - 1 0 0 - - 1 0 0 0 1 0 - 0 1 - 2 1 1 - 0 0 - - 0 - - - 0 0 0 1 1 0 0 1 0 - - - - - - 1 - - - - 1 0 0 0 - 0 - - 0 0 - - 1 1 0 1 0 1 0 1 0 - 0 0 - - - 1 0 0 - - 0 0 1 1 2 1 0 1 1 1 1 1 0 1 0 1 0 1 1 1\nActinarctus_Heterotardigrada 1 1 0 0 - - - 0 0 1 1 0 0 1 0 0 1 0 1 0 0 1 0 0 0 1 1 0 0 0 - 0 - 0 - 1 0 - - 0 - - - 0 0 0 0 - - - - 1 1 0 0 - 1 0 0 ? ? - - - 0 0 0 - 0 - - 0 0 - - 0 0 1 1 0 3 1 0 0 - 0 0 - - - 0 0 1 1 - 1 1 1 0 0 1 1 0 1 1 1 0 1 1 ? - 1 0 0 -\nHalobiotus_Eutardigrada 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 1 0 1 0 0 1 0 0 0 1 1 0 0 0 - 0 - 0 - 2 0 - - 0 - - - 0 0 0 0 - - - - 1 0 - - - - - 0 ? ? - - - 0 0 0 - 0 - - 0 0 - - 0 0 1 1 1 1 1 0 0 - 0 0 - - - 0 0 1 1 - 1 1 1 0 0 1 1 0 1 1 1 0 1 1 1 - 1 0 0 -\nSiberian_Orsten_tardigrade 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? ? ? ? ? ? [0-] ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 0 - - - - ? ? ? - - ? ? ? ? ? - - - ? 0 0 - 0 - - 0 0 - - 0 0 0 1 1 1 1 0 0 - 0 0 - - - 0 0 1 [12] - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nMegadictyon 1 0 - - - - 0 ? 0 0 ? ? ? 1 ? ? ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 0 - - 1 0 - 1 0 0 0 1 1 ? 0 ? 0 - - - - - - ? - - - - 0 1 0 0 - 0 - - 0 1 1 - 0 0 0 ? ? ? ? 0 0 - 0 0 - - - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nJianshanopodia 1 0 - - - - 0 0 0 0 0 - 0 1 ? ? 1 0 1 1 0 0 - - 0 1 0 - 0 0 - 0 - 0 - 0 ? ? ? 1 0 - ? 0 0 0 1 1 ? 1 0 0 - - - - - - ? - - - - 0 1 0 0 - 0 - - 0 1 1 - 0 0 0 ? ? ? ? 0 0 - 0 0 - - - 0 0 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHadranax 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ? 0 ? - ? ? ? - ? ? ? ? ? ? ? ? 0 0 0 1 1 ? 1 0 1 1 0 0 - 0 - ? 3 - - 0 1 ? 0 0 - 0 - - 0 0 - - 1 0 ? ? ? ? ? 0 0 - 0 0 - - - ? 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nKerygmachela 1 0 - - - - 0 0 0 0 0 - 0 1 0 0 ? 0 ? ? ? ? ? ? 0 1 0 - 0 0 - 0 - 0 - 0 1 0 - 1 0 - 1 0 0 0 1 1 0 1 0 1 1 0 0 - 0 - ? 3 - ? 0 1 1 0 1 0 0 1 0 0 0 - - 0 0 0 0 0 - - 0 0 0 0 0 - - - 0 0 1 0 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 0 ? ? ?\nPambdelurion 1 0 - - - - 0 0 0 0 1 0 0 1 0 0 ? 0 ? ? ? 0 - - 0 1 0 - 0 0 - 0 - 0 - 1 1 0 - 1 0 - ? 0 0 0 1 ? ? ? ? ? ? ? ? ? ? ? ? - - ? - ? 1 0 1 0 0 1 0 0 0 - - 0 0 0 0 0 - - 0 0 0 0 0 - - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 1 0 -\nOpabinia 1 0 - - - - 0 0 0 0 2 1 0 ? ? ? ? 0 ? ? ? 1 1 1 0 1 0 - 0 0 - 0 - 0 - 1 1 1 0 1 0 - 1 1 0 0 0 - - - - 0 - - - - - - - - - - - - 1 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 0 0 0 0 - - - ? 1 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAnomalocaris_canadensis 1 1 0 1 0 0 0 0 ? 0 1 1 0 1 1 1 1 0 ? ? ? 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 1 0 - 1 1 0 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - 1 0 1 0 0 0 1 0 0 - - 0 0 0 0 0 - - 0 1 1 0 1 - - - 0 1 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? ?\nPeytoia_nathorsti 1 1 0 1 1 0 0 0 ? 0 1 1 0 1 1 1 ? 0 ? ? ? 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 1 0 - 1 0 - 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - 1 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 1 1 0 1 - - - 0 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nHurdia_victoria 1 1 0 1 1 0 1 0 ? 0 1 1 0 1 1 1 1 0 1 1 0 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 ? ? ? 1 0 - 1 ? 0 0 0 - - - - 0 - - - - - - - - - - - - ? 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 1 0 0 1 - - - 0 0 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAegirocassis_benmoulae 1 1 0 1 1 0 1 0 ? 0 ? 1 0 ? ? ? ? ? ? ? ? ? ? ? 0 1 1 1 1 0 - 0 - 0 - 1 ? ? ? 1 0 - 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - ? 0 1 0 0 1 1 0 0 - - 0 0 0 0 0 - - 0 1 1 0 ? - - - 0 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nLyrarapax_unguispinus 1 1 0 1 0 0 0 0 ? 0 1 1 0 1 ? ? ? ? ? ? ? 1 1 1 0 1 1 1 1 0 - 0 - 0 - 1 1 0 - 1 0 - 1 0 0 0 0 - - - - 0 - - - - - - - - - - - - 0 0 1 0 0 0 1 0 0 - - 0 0 0 0 0 - - 0 ? 1 1 1 - - - 0 1 1 0 1 ? ? 1 0 0 1 1 ? ? ? ? ? ? ? ? ? 1 ? ? ?\nSchinderhannes 1 ? ? ? ? ? ? 0 ? 0 1 1 0 1 ? ? ? ? ? ? ? 1 1 1 0 1 1 1 ? ? - 0 - 0 - 1 1 0 - 1 1 0 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? - - - 0 0 1 0 1 ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nFuxianhuia 1 1 0 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 1 1 1 1 1 - - - - 1 1 0 - - - - - - - - - - - - - - - - - - 0 - 1 1 1 - 0 0 - - - - - 0 0 0 - - 0 0 - 0 0 - - 1 - 0 1 0 1 ? ? 1 2 1 1 1 ? ? ? ? ? ? ? ? 1 1 0 0 -\nChengjiangocaris 1 1 0 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 1 1 1 1 1 - - - - 1 1 0 - - - - - - - - - - - - - - - - - - 0 - 1 1 1 - 0 0 - - - - - 0 0 0 - - 0 0 - 0 0 - - 1 - 0 1 0 1 ? 1 ? ? ? ? ? 0 ? 1 ? ? ? ? ? ? ? ? ? ?\nLeanchoilia 1 1 1 0 - - - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 1 - - - 0 1 1 1 1 - - - - 1 1 ? - - - - - - - - - - - - 0 - - - - - 1 - 1 1 1 - 0 1 - - - - - 0 1 0 [02] ? 0 0 - 0 0 - - 0 - 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ?\nAlalcomenaeus 1 1 1 0 - - - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 ? ? - 1 1 - - - 0 1 1 1 1 - - - - 1 1 ? - - - - - - - - - - - - 0 - - - - - 1 - 1 1 1 - 0 1 - - - - - 0 1 0 0 - 0 0 - 0 0 - - 0 - 0 0 - - ? 1 1 2 1 1 1 0 0 0 - - ? ? ? ? ? ? ? ?\nMisszhouia_longicaudata 1 1 1 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 0 1 1 1 1 - - - - 1 1 1 - - - - - - - - - - - - 0 - - - - - 1 - 1 2 1 - 0 1 - - - - - 0 1 0 0 - 0 0 - 0 0 - - 0 - 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? ?\nKuamaia_lata 1 1 1 1 0 1 - 0 - - 2 1 0 0 - - ? 0 ? ? ? 1 1 1 1 1 1 0 - 1 0 - - - 0 1 1 1 1 - - - - 1 1 1 - - - - - - - - - - - - 0 - - - - - 1 - 1 2 1 - 0 1 - - - - - 0 1 0 2 1 0 0 - 0 0 - - 0 - 0 0 - - ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? ? 1 ? ? ?\n;\nxpiwe=;"

# Write to a file with:
# WriteTntCharacters(Lobo.phy, "example_file.tnt")